Chore : 로그 수정 및 RestController분리
