+++ src/main/java/com/sysone/ddogdog/customer/hotel/controller/HotelController.java
@@ -2,7 +2,6 @@ package com.sysone.ddogdog.customer.hotel.controller;
 
 import com.sysone.ddogdog.customer.hotel.model.ResponseHotelDTO;
 import com.sysone.ddogdog.customer.hotel.service.HotelService;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -11,6 +10,8 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -21,7 +22,7 @@ public class HotelController {
 
     @GetMapping("/v1/customers/getDataList")
     public ResponseEntity<List<String>> getDataList(@RequestParam String searchKeyword) {
-        List<String> dataList= hotelService.getDataList(searchKeyword);
+        List<String> dataList = hotelService.getDataList(searchKeyword);
         return ResponseEntity.ok(dataList);
     }
 
@@ -31,8 +32,7 @@ public class HotelController {
                                                             @RequestParam(defaultValue = "1") int page,
                                                             @RequestParam(defaultValue = "3") int size) {
         Page<ResponseHotelDTO> hotels = hotelService.getHotelsByKeywordAndDatesWithPagination(keyword, startDate,
-                endDate,page,size);
-        log.info("페이지네이션 결과 반환 -> "+hotels.toString());
+                endDate, page, size);
         return ResponseEntity.ok(hotels);
     }
 }
