+++ src/main/java/com/sysone/ddogdog/customer/reservation/controller/ReservationController.java
@@ -5,9 +5,7 @@ import com.sysone.ddogdog.customer.reservation.model.RequestReservationDTO;
 import com.sysone.ddogdog.customer.reservation.model.ResponseMostReservationHotelDTO;
 import com.sysone.ddogdog.customer.reservation.model.ResponseReservationStatsDTO;
 import com.sysone.ddogdog.customer.reservation.service.ReservationService;
-import com.sysone.ddogdog.common.exception.NoAvailableRoomsException;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -16,16 +14,15 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequestMapping("/v1/customers/reservation")
 @RequiredArgsConstructor
-@Slf4j
 public class ReservationController {
 
     private final ReservationService reservationService;
 
     @PostMapping
     public ResponseEntity<String> saveReserve(@AuthenticationPrincipal PrincipalDetails user,
-        @RequestBody RequestReservationDTO dto) {
-            reservationService.saveReserve(user,dto);
-            return new ResponseEntity<>("예약에 성공했습니다", HttpStatus.CREATED);
+                                              @RequestBody RequestReservationDTO dto) {
+        reservationService.saveReserve(user, dto);
+        return new ResponseEntity<>("예약에 성공했습니다", HttpStatus.CREATED);
     }
 
     @PatchMapping("/{id}")
@@ -35,13 +32,13 @@ public class ReservationController {
     }
 
     @GetMapping("/stats")
-    public ResponseEntity<ResponseReservationStatsDTO> findReservationStats(@AuthenticationPrincipal PrincipalDetails user){
+    public ResponseEntity<ResponseReservationStatsDTO> findReservationStats(@AuthenticationPrincipal PrincipalDetails user) {
         ResponseReservationStatsDTO statsDTO = reservationService.findReservationStatsByCustomerId(user.getUsername());
         return ResponseEntity.ok(statsDTO);
     }
 
     @GetMapping("/mostHotel")
-    public ResponseEntity<ResponseMostReservationHotelDTO> findMostReservationHotelByCustomerId(@AuthenticationPrincipal PrincipalDetails user){
+    public ResponseEntity<ResponseMostReservationHotelDTO> findMostReservationHotelByCustomerId(@AuthenticationPrincipal PrincipalDetails user) {
         ResponseMostReservationHotelDTO mostHotelDTO = reservationService.findMostReservationHotelByCustomerId(user.getUsername());
         return ResponseEntity.ok(mostHotelDTO);
     }
